requirements.txt
pillow
matplotlib
torch
torchvision
numpy
tqdm
scikit-image
streamlit
